Refactoring: renaming the tasks and removing the unused bmp sensor files
