--- AvionicsSoftware-AtollicProject/Inc/tasks/cli.h
@@ -1,138 +1,0 @@
-#ifndef XTRACT_H
-#define XTRACT_H
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// UMSATS 2018-2020
-//
-// Repository:
-//  UMSATS>Avionics-2019
-//
-// File Description:
-//  xtraxt.h UART CLI utility to pull data off of STM32 flash memory
-//
-// History
-// 2019-02-15 by Eric Kapilik
-// - Created.
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// INCLUDES
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include "forward_declarations.h"
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// DEFINITIONS AND MACROS
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// ENUMS AND ENUM TYPEDEFS
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-typedef enum{
-
-	MAIN_MENU,
-	READ_MENU,
-	CONFIG_MENU,
-	EMATCH_MENU,
-	MEM_MENU,
-	SAVE_MENU
-
-} menuState_t;
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// STRUCTS AND STRUCT TYPEDEFS
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-typedef struct{
-
-	UART_HandleTypeDef *huart;
-	Flash *flash;
-	configData_t *flightCompConfig;
-	TaskHandle_t startupTaskHandle;
-}	xtractParams;
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// TYPEDEFS
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// CONSTANTS
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// FUNCTION PROTOTYPES
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  create task to read & handle UART connection
-//
-// Returns:
-//   VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void vTask_CLI(void *pvParameters);
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  handle command passed in
-//		- does not create command parameter deep copy
-//		- will call one of the other commands declared below
-//
-// Returns:
-//  VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void handle_command(char* command,xtractParams * params,menuState_t* state);
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  display introduction about xtract program
-//
-// Returns:
-//  VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void intro(UART_HandleTypeDef * uart); //display on start up
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  display help menu, including short description of all available commands
-//
-// Returns:
-//  Enter description of return values (if any).
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void help(UART_HandleTypeDef * uart); //display help menu
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  read data from flash memory - CURRENTLY DUMMY FUNCTION
-//
-// Returns:
-//  VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void read(xtractParams * params);
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  starts a timer that prints when it is done
-//
-// Returns:
-//  VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-//void start(void);
-
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  menu for ematch options.
-//
-// Returns:
-//  VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void ematch(char* command, xtractParams * params);
-
-
-void memory_menu(char* command, xtractParams * params);
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-// Description:
-//  menu for setting flight computer options.
-//
-// Returns:
-//  VOID
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void configure(char* command,xtractParams * params);
-
-#endif // XTRACT_H
