+++ AvionicsSoftware-AtollicProject/Src/tasks/sensors/imu_sensor.c
@@ -18,7 +18,7 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include "tasks/sensorAG.h"
+#include "tasks/sensors/imu_sensor.h"
 #include "bmi088.h"
 #include "SPI.h"
 #include "hardwareDefs.h"
