+++ AvionicsSoftware-AtollicProject/Src/tasks/startup.c
@@ -15,7 +15,7 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include "tasks/startupTask.h"
+#include "tasks/startup.h"
 
 #include "cmsis_os.h"
 #include "configuration.h"
